DYEING CLOTHES! (#342)
You can now dye a select few types of clothes, with the possibility of
more clothes in the future if people see it fit..!

No longer will we have six different ways to craft a colored robe - no
longer will you have to suffer with a horrible mismatch of colors as a
mage - no longer will you have to wear BROWN PEASANT PANTS as a noble -
now you can be STYLISH with just a bit of effort!

![image](https://github.com/Rotwood-Vale/Ratwood-Keep/assets/11761309/9197a5e4-2acf-4d9e-87ed-0e891ffa4af5)

Basically ported colormate from desertrose and made it into a 'dyeing
station'. No good at spriting so the sprite is just a stone bin.

https://github.com/Rotwood-Vale/Ratwood-Keep/assets/11761309/a34da65a-1b11-40a1-a341-4f766ce2e5ac

Also comes with the squashing of all robe craftable color subtypes into
one AND dr2's rgb matrices helpers

---------

Co-authored-by: Azarak <azarak10@gmail.com>