Fixes a bug where Lupians are forced to be hermaphrodites.
This bug was caused by listing all of the sexual organs in the list of organs the species is supposed to start with, which adds them with complete disregard to the player's preferences.

For future reference for coders making new species, sexual organs only need to be listed in the customization list, as that system handles whether or not to add the organs themselves to a character.