Katar: The unarmed skill weapon (#403)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

It adds a new weapon, the Katar! a blade that rests above the fist, it
has a thrust and cut attack, fast speed, low damage, and uses the
unarmed skill. It can be crafted for 1 steel and smelted for 1 steel
too.


![kataringame](https://github.com/Rotwood-Vale/Ratwood-Keep/assets/173851742/f27497e5-8889-4120-b682-d7cae2d60339)
Photo of the katar sprite

![kataringame2](https://github.com/Rotwood-Vale/Ratwood-Keep/assets/173851742/f531fa13-b8b7-48b9-9420-bcbaeb626e4f)
Photo of how the katar looks while being held.
## Why It's Good For The Game


It adds more variety to the amount of choices for weapons in the game,
adds a choice for unarmed skill weapons (stabbing and cutting with the
katar is almost exactly the same movement and technique as throwing a
punch), a skill that was rarely to almost never used.
It's also pretty fuckin cool weapon.

---------

Co-authored-by: LucaNotLucas <lucagiachero08@gmail.com>