Job greeting text (#507)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

Greets jobs with their Title and their Tutorial

Rewrites Pilgrim tutorial text to be more IC and less OOC

![image](https://github.com/user-attachments/assets/354dd813-4829-4104-9e39-7d6e41bcbe96)

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

Gives people an idea who they are and what they're supposed to do if
they dont pay too much attention to the job selection menu

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->