Some helmet rebalancing
Hey. Did you know the visored sallet mysteriously has a not-infuriating vision cone, out of all the face-protecting helmets in the game? This is supposedly because it's a "light" helmet, with less protection than it's "heavy" siblings like the pigface bascinet. Wanna know something else that's interesting?

Besides durability, these helmets actually all have the same protective values.

The only difference is that the "heavy" helmets with horrendous vision cones have double the durability at 400, while most of the non-face-protecting ones have 200. This, frankly, doesn't make any sense, and as good face-protecting helmets are virtually mandatory for combat-focused roles alongside neck protection, this forces players to play with 25% vision radius because... "skill issue"? This increases the "default" vision radius of masked helms to 50% as a quality of life thing while preserving the balance of the helmets.

"But cofi!" you may ask, "What will become of the visored sallet, since it's one redeeming quality was being a visored helm with shit durability but face/eye protection!" The answer is I've given it "heavy" durability, making it an identical and equally viable choice.

"But cofi!" you may ask, "30 or 40% seems more reasonable!" And I would say, unfortunately, the code seems to only care for "right" and "left" and "behind", and coding in entirely new vision values is far beyond my ability. If anyone can do it, please, stand up. I've tested this, it looks good, and it feels good. The protective values are all the same as before (that is to say, there was no difference between the two I think.)