Health potion buff (#533)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
This requested PR allows healing potions to repair wounds such as
fractured ribs, aortic dissections, etc. the potion heals 2 woundHP per
.5 units imbibed. If you are fucked up you will need several units.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
This is the requested alternative to miracles or surgeons without
removing either. This also heals over time so it cannot be used to
become the god of combat. This was requested by the host and they were
kept in the full loop of development.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->