Chat to 1000 (actually 200)! (#543)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

Simply increases the chat's maximum lines from 100 to 200.

## Why It's Good For The Game

Currently, with how short the chat box is, if you spend about thirty
seconds in a busy bar or shift examine someone two or three times by
accident, anything said before disappears. This simply increases it to a
more reasonable size.