Vampires can only slurp 9000 vitae from one person (#593)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
Fix to unhealthy bloodbag healing, and to discourage keeping one captive
for 2 hours
Better than any bandaid like
https://github.com/Rotwood-Vale/Ratwood-Keep/pull/589 I think

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->