Mirror Description (#576)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->
Changes mirror to have a fun little "It's me!" description when
examined. Also a test PR because I have barely any idea how to use
GitHub or program. Learning is fun!

## Why It's Good For The Game

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->

Self explanatory.

---------

Co-authored-by: Gyrandola <pasta.frollagg@gmail.com>