Makes frills Toggleable(?)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may not be viewable. -->

## About The Pull Request
Simply changed the "ORGAN_SLOT_FRILLS = /obj/item/organ/frills/lizard," and the "/datum/customizer/organ/frills/lizard,", to be /anthro instead, it's functionally no different other than not being toggleable. I may make a PR for Sissian to have the same change as again, it's functionally no different outside of a quality of life change.
<!-- Describe The Pull Request. Please be sure every change is documented or this can delay review and even discourage maintainers from merging your PR! -->

## Why It's Good For The Game
Dragons vary greatly in terms of their descriptions based on culture. Instead of adding various different races to accommodate this, a minimalistic change that let's you not be some kind of fish-dragon with frills is a good quality of life change for those wanting to play the race. This was a request by two people, maybe more on the discord to change things.
<!-- Argue for the merits of your changes and how they benefit the game, especially if they are controversial and/or far reaching. If you can't actually explain WHY what you are doing will improve the game, then it probably isn't good for the game in the first place. -->