Re-palleting the bog
Adding some slightly brighter colors to the world. The bog is already filled with deciduous trees, and the only dead trees that exist are the ones around the Keep, so I'm not really sure why the rest of the world has to be so ugly. Might be a nice detail to keep the original ugly colors to be around Rockhill only, but this is just a recoloring of the icons.