Add Sea jacket to crafting (#643)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

Added the sea jacket, which is equivalent to the leather vest (with the
exception that you can only use it on the armor slot), currently only
legendary fishers(very low chance to get it) and grabbers(disabled) get
them, which kinda stuck.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

Axian Stuffs, we got the bicorn and the bandana, let's add that aswell
for peoples to do their gimmick

![image](https://github.com/user-attachments/assets/14a44bb2-b924-42d6-b601-9507b4e33dd5)

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->