Un-powercreep smith (#641)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
removes a lot of combat skills from smith, rebalances their stats and
buffs their smithing skills to what it was before a sexist powercreeper
touched smith (see
https://github.com/Blackstone-SS13/BLACKSTONE/commit/95b63d905b1b5593bf933af46ca7234e4777a3f5#diff-dd504dad939397b8e041470a3dd4b088e5b30a3795212ed21ebdab5efa64aa10)
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
This lets smiths do their job better and lets less strong races be able
to smith better see notes for in depth explanation on how the changes
are also realistic
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->