Hellspeak changes and fixes (#636)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
Fixes Hellspeak not having a font and changes the font it uses for a
more "demonic" one.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Fixes a bug and gives the language a better looking font in my eyes.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->

![proof1](https://github.com/user-attachments/assets/de2e0528-8571-4cf1-995b-03484200f55a)

![Proof2](https://github.com/user-attachments/assets/44f05447-d187-4d2b-a475-218b7cd47456)