All PQ changes will fail without a reason (#667)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
Makes AdjustPQ fail with a messsage if you either enter a blank reason,
or use the Cancel button.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Without this PR, if you cancel when it prompts you to enter a reason, it
will apply the PQ adjustment anyways, just without a message. I hate
this jank so fucking much lmaooo
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->

Co-authored-by: zoktiik <zoktiik@users.noreply.github.com>