Makes the Dying Station more immersive (#609)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
Now it's called "dye bin"
Now it needs 1 small log to build
Now it can be grabbed, pulled, thrown, manipulated around, and destroyed
easier
Now it requires juicing berries into it to grant 3 uses of dying clothes
Dying clothes now requires a small do_after() timer of 3 SECONDS
Made all sorts of descriptions of the bin and usages of it much more
fitting and immersive


![dreammaker_ndXew3yeH0](https://github.com/user-attachments/assets/afb96014-2834-4987-81ed-1ea9c211753d)



<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->