Manual Genital Arousal (#635)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

This PR adds the ability to manually set the erection state of male
genitals via sexcon, next to speed and force, which is set to natural
(matching arousal level) by default. This option is hidden if the player
lacks the suitable organ (penis).


![Untitled](https://github.com/user-attachments/assets/c29fdc41-9039-4a3c-8332-6801baf9b4b0)

## Why It's Good For The Game

With the welcome addition of genital states, the ability to set it
manually was conspiciously lacking, requiring players to use mechanics
even if they dislike using the system to make their sprite match the
described state. This PR will allow emote-preferring players to have
their sprite match the scene without using mechanical sex. It also
allows for arousal due to non-physical causes (i.e. morning wood or
simply seeing someone naked), or remaining unerect despite of physical
contact (such as being a slow grower, unattracted to your partner or
simply erectile dysfunction.)

While many scenes will not need this, many others will benefit from it,
and the addition of a setting that can be ignored if irrelevant is a
non-inconvenience.