Adds Mouth-to-Mouth to combat oxyloss. (#614)
## About The Pull Request

- Adds a Mouth-to-Mouth "surgery" to combat the oxyloss most commonly
caused by bleeding out.

Works like this:
- Target mouth on a person on the floor with touch intent on weak. If
they have oxy-damage, you will start breathing air into their lungs. It
repeats on success, if you fail, you need to click again. So you really
need to be paying attention to it.

I tinkered with the numbers a little, with 0 medicine skill you have a
~40% to ~60% failure chance each time, barely enough to keep people in
the round until real help arrives. If you have a needle and water
available right as they pass out, you might get lucky and save them all
by yourself. The chance is low, but it is there. If not, better start
shouting for people to bring those things while you struggle against the
oxyloss.

## Why It's Good For The Game

It is frustrating as hell to stitch someone up completely, just for them
to expire when the last wound was stitched. This PR adds a way to delay
the inevitable, and as long as the survivors are able to stitch wounds
and bring water, people can be kept in the round more easily.

- Passing out from bloodloss is no longer a death sentence, if care can
be provided in time.
- Keeps people in the round. At least people that have friends.
- Antags will have a slightly easier time to stay in the round if their
comrades win the fight, and they can sometimes capture people that would
just have died a stupid death before.
- Vampires accidentally critting when biting won't mean certain death
for their victim anymore, if the vampire cares enough to provide aid.