Changes Pick combat skill to maces from mining (#608)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
A legendary miner should not be a legendary combatant with a pick...
Miners now get a mace skill instead of axe skill (sets them apart from
lumberjacks, and Pick pierces armor kind of fitting as a "mace")

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->