Prince/Princess Stat/Skill Buff (#566)
## About The Pull Request

This PR changes the Statarray of the prince/princess to be slightly less
inept, to encourage doing anything but ERP in the keep. Nobility used to
be full of gloryseekers and many young princes especially died seeking
glory, which is a bit hard to pull off when your stat array contains a
-2 STR.
Their new stat array should be a total of +4, from +0.
The description is also changed to somewhat reflect a young adult and
not the implication of a teenager.

I've alternatively considered removing the gendered stat and skill
arrays and giving them subclasses, but I feel like this may be one of
the few places where we can be a bit more traditionalist and keep the
gender split to a degree. If we want to get rid of that too, then that's
fine.

## Why It's Good For The Game
Increase consistency with the nobility being at slightly competent (not
very but not atrocious).
Encourage doing literally anything from the princes/princess like
leading men into battle or dueling.
Move away from the princes/princess = teenage deadweight implications.