Royal Guard Update + Leper Flaw  + New Armor + Crossbow Fix (#249)
* Royal Guard Update

Updates Royal Guard armour to look more fancy.

* Leper Flaw

Adds Leprosy as a flaw. Currently doesn't do much other than sometimes cause limbs to spasm and gives notifications regarding sickness. As it stands this is mostly a roleplay Flaw as it doesn't really have the main symptoms and flaws the actual one would have, but I still made it for people to build off of.

* New Gear

Adds barrel armour, a new helmet for Eoran templars, a set of Pysdonian robes and a Pysdonian hood, and a new cosmetic version of the bascinet with a barred visor.

* New headgear

Adds new headgear.

* Crossbow fix

Fixes the current bug that prevents the crossbow from being uses properly.

* Beak Helmet recipe

* Beak Helmet Recipe Fix.dm

* Update matchlock.dm

* Update matchlock.dm

Readds flintlock intents