Makes Cloaks Show Tails Unless Hood Is Up (#711)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

This is a simple change that makes it so cloaks no longer cover up
tails. The tail is now hidden when the cloak's hood is drawn, allowing
for optimal sneaking around still.

## Why It's Good For The Game

This feature has been widely requested - the number of players who
discard cloaks simply because it obscures the tail is quite high. While
hiding the tail helps sneaking around in it, this is not a concern when
you have your face exposed for all to see. This gives a best of both
worlds, where your character can look non-ugly while wearing the cloak
and still obscuring them when you want to hide your identity with the
hood.