You can heat/cook stuff over a forge now, tell temperature with a thermoscope, grind a lot more stuff, and changes some recipes that were technically craftable to be more lorefriendly. (#542)
A few chemical recipes turned out to be craftable, albeit it was really
fucking hard to make them (fentanyl from heating space drugs and
multiver from ash and salt): They have been given appropriate theme
names and fentanyl actually does something.

It was also incredibly weird but possible trying to heat any liquid
source to high temperatures using non-hearth methods, so I added the
ability to heat reagents to greater temps on top of forges.

I also changed 'glass pot' to 'metal pot' because it was meant to be a
metal pot - you literally craft it out of metal. The guy who just edited
it assumed everything that started with /glass/ in the code meant it had
to be made of glass, ignoring shit like the wooden buckets... Check the
code comments, they're very stupid.

You could technically make fentanyl by heating grinded dried swampweed.
You can still do that but the drug has been renamed to 'Goblin' and
gives you the appropriate moodlet (works better on the stress refactor
tm).

You can now grind a lot more things that I meant to make grindable
originally but didn't have the time to add. Ores and ingots and organs
and such.

Most importantly, you can now make a thermoscope to detect the
temperature of the reagents in a container. It's made by mining for
Cinnabar, a relatively uncommon ore, grinding it into mercury, and
crafting it with a glass bottle. It's the lowest level alchemy craft and
gives level 0 high int people an in to start getting alchemy.

You can only use the thermoscope if you are at least a level 1
alchemist, and you can only use it with actual precision if you are a
level 4 alchemist. It uses 'room temperature' to determine the heat so
even then it's a bit hard to read given how lowtech it is.

---------

Co-authored-by: Azarak <azarak10@gmail.com>