Some more map changes (#697)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
One of the homes near the town square is now a dining area
Added a lock to the doctor's chest
Added two deadbolt shacks in the basin, outside of the town, both hidden
by some trees

Also adds windows with curtains and refactors their code to be better

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->