Nix the Nudist class (#695)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
Disables the Nudist from being a pickable class.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
The existence of nudist as a class causes far more animosity between
players than is necessary, due to the nudist trait mechanically blocking
you from wearing clothes, thus giving you an entirely OOC excuse to
argue about why you should avoid consequences for being nude. It's far
better as something you deliberately roleplay when nudism is widely
shunned.

A lot of the roleplay happens from behind the bog gates, and being
blocked due a mechanical barrier preventing you from becoming more
socially acceptable is not a good thing. Especially when Nudist in its
current state doesn't do anything to add to the gameplay of being out in
the bog, seeing as you're not even a survivalist. Just get rid of it
instead of forcing players to engage in this sort of conflict.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->