Butchery changes (#683)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
The Butchery now starts locked, also makes the animal pen less dark.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Since Butcherer starts with keys to their shop, it makes sense for it to
start locked, and the animal pens become pitch black during the night,
so some lighting should be good.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->

![Captura](https://github.com/user-attachments/assets/b98b0c9e-32df-48f7-b0ae-dcf79c7c26c6)