Remove hammer req for crafting improvised weapons (#696)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
Requested by ezekyal on Discord. Simply removes the hammer as a tool
requirement for crafting the goedendag, improvised billhook, and peasant
war flail.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
From what it looks like: the need for a hammer to craft these makes them
more difficult to make than would be necessary for an improvised weapon.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->