Bandit look fix (#295)
* Creates a basic catch clause for bandit examines

Bandits, on creation, didn't get a job (null) which means they didn't have a title. FIXES this, and also makes it so there's a catch in the examine so when you look at them it works.

* Adds a comment

that's it.

* Fixes based on Deltafire's posts

I actually didn't think about else if. I appreciate people double checking me.