Adds more items to the bandit sacrifice list and refactor (zoktiik) (#720)
## About The Pull Request

Adds:
riddle of steel
emerald choker
desert rider amulet
merchant amulet
rockhill crown
paintings
-
To the list of things that can be sacrificed at the bandit idol.

## Why It's Good For The Game

I think the only things I need to justify here are the crown and the
paintings.
If the bandits actually manage to get the crown, and feed it to the
idol. Well. I guess bandits win. Besides killing the local government
off, that's peak matthios right there with sacrificing the object of
their hatred. It just makes sense to let them destroy it. If the town
screwed up that badly they deserve it.

The paintings are actually kinda valuable hovering around 30-40 mammon
give or take. And it's a trope with raiders stealing paintings in media.
So why not?

---------

Co-authored-by: zoktiik <zoktiik@users.noreply.github.com>