Fixes kneestinger midcombat power (#649)
## About The Pull Request
Fixes kneestingers being the most overpowered combat mechanic in the
game.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Kneestingers are blatantly overpowered, they needed a nerf, and everyone
knew this.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->
The enjoyment of the many should not be sacrificed for the fragging of
the few.